+++ src/app/post/[slug]/loading.tsx
@@ -1,4 +1,4 @@
-import PostPageSkeleton from '@/features/post/public/components/PostPageSkeleton';
+import PostPageSkeleton from '@/features/post/public/postDetail/components/PostPageSkeleton';
 
 export default function Loading() {
   return <PostPageSkeleton />;
