+++ src/app/post/[slug]/page.tsx
@@ -1,11 +1,11 @@
 'use client';
 
-import PostTitle from '@/features/post/public/components/PostTitle';
-import PostInformation from '@/features/post/public/components/PostInformation';
+import PostTitle from '@/features/post/public/postDetail/components/PostTitle';
+import PostInformation from '@/features/post/public/postDetail/components/PostInformation';
 import { ProjectPublicInfo } from '@/features/project/public/components/ProjectPublicInfo';
-import PostIntroduction from '@/features/post/public/components/PostIntroduction';
+import PostIntroduction from '@/features/post/public/postDetail/components/PostIntroduction';
 import ApplyProject from '@/features/projectApply/auth/components/ApplyProject';
-import { usePostPublicInfo } from '@/features/post/public/service/getPostPublicInfo';
+import { usePostPublicInfo } from '@/features/post/public/postDetail/service/getPostPublicInfo';
 import { numStrToBigInt } from '@/utils/common';
 import { useProjectPublicInfo } from '@/features/project/public/service/getProjectPublicInfo';
