+++ src/features/post/public/postDetail/components/PostInformation.tsx
